requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.8
h11==0.14.0
idna==3.10
kombu==5.4.2
lxml==5.3.1
memory-profiler==0.61.0
motor==3.7.0
opcua==0.98.13
paho-mqtt==2.1.0
prompt_toolkit==3.0.50
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pymodbus==3.8.6
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
